requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
bleach=1.5.0=py36_0
ca-certificates=2017.08.26=h1d4fec5_0
cffi=1.9.1=py36_0
click=6.7=py36_0
cudatoolkit=8.0=3
cudnn=7.0.5=cuda8.0_0
cycler=0.10.0=py36_0
dbus=1.10.22=h3b5a359_0
decorator=4.0.11=py36_0
entrypoints=0.2.2=py36_1
expat=2.1.0=0
flask=0.12.2=py36_0
fontconfig=2.12.4=h88586e7_1
freetype=2.8=hab7d2ae_1
glib=2.53.6=h5d9569c_2
gst-plugins-base=1.12.4=h33fb286_0
gstreamer=1.12.4=hb53b477_0
hdf5=1.8.17=1
html5lib=0.999=py36_0
icu=58.2=h9c2bf20_1
intel-openmp=2018.0.0=hc7b2577_8
ipykernel=4.6.1=py36_0
ipython=6.2.1=py36h88c514a_1
ipython_genutils=0.2.0=py36_0
ipywidgets=6.0.0=py36_0
itsdangerous=0.24=py36_0
jbig=2.1=0
jedi=0.11.1=py36_0
jinja2=2.9.6=py36_0
jpeg=9b=0
jsonschema=2.6.0=py36_0
jupyter=1.0.0=py36_3
jupyter_client=5.0.1=py36_0
jupyter_console=5.1.0=py36_0
jupyter_core=4.3.0=py36_0
libedit=3.1=heed3624_0
libffi=3.2.1=1
libgcc=5.2.0=0
libgcc-ng=7.2.0=h7cc24e2_2
libgfortran-ng=7.2.0=h9f7466a_2
libiconv=1.14=0
libpng=1.6.34=hb9fc6fc_0
libsodium=1.0.10=0
libstdcxx-ng=7.2.0=h7a57d05_2
libtiff=4.0.9=h28f6b97_0
libxcb=1.12=1
libxml2=2.9.4=0
markupsafe=0.23=py36_2
matplotlib=2.1.2=py36h0e671d2_0
memory_profiler=0.47=py36hfa32046_0
mistune=0.7.4=py36_0
mkl=2018.0.1=h19d6760_4
nbconvert=5.1.1=py36_0
nbformat=4.3.0=py36_0
ncurses=6.0=h9df7e31_2
notebook=5.0.0=py36_0
numpy=1.14.1=py36h3dfced4_1
olefile=0.44=py36_0
openssl=1.0.2n=hb7f436b_0
pandas=0.21.1=py36h20b78c2_0
pandocfilters=1.4.1=py36_0
parso=0.1.1=py36h35f843b_0
path.py=10.3.1=py36_0
patsy=0.4.1=py36ha3be15e_0
pcre=8.39=1
pexpect=4.2.1=py36_0
pickleshare=0.7.4=py36_0
pillow=5.0.0=py36h3deb7b8_0
pip=9.0.1=py36_1
prompt_toolkit=1.0.14=py36_0
psutil=5.2.1=py36_0
ptyprocess=0.5.1=py36_0
pycparser=2.17=py36_0
pygments=2.2.0=py36_0
pyparsing=2.2.0=py36_0
pyqt=5.6.0=py36_2
python=3.6.4=hc3d631a_1
python-dateutil=2.6.0=py36_0
pytorch=0.3.1=py36_cuda8.0.61_cudnn7.0.5_2
pytz=2017.2=py36_0
pyyaml=3.12=py36_0
pyzmq=16.0.2=py36_0
qt=5.6.2=h974d657_12
qtconsole=4.3.0=py36_0
readline=7.0=ha6073c6_4
scipy=1.0.0=py36hbf646e7_0
setuptools=27.2.0=py36_0
simplegeneric=0.8.1=py36_1
sip=4.18=py36_0
six=1.10.0=py36_0
sqlite=3.22.0=h1bed415_0
statsmodels=0.8.0=py36h8533d0b_0
terminado=0.6=py36_0
testpath=0.3=py36_0
tk=8.6.7=hc745277_3
tornado=4.5.1=py36_0
traitlets=4.3.2=py36_0
wcwidth=0.1.7=py36_0
werkzeug=0.12.1=py36_0
wheel=0.29.0=py36_0
widgetsnbextension=2.0.0=py36_0
xz=5.2.3=h55aa19d_2
yaml=0.1.6=0
zeromq=4.1.5=0
zlib=1.2.11=ha838bed_2
